Move toogle log visibility function to model; add if condition to journey post view

+++ journeys/views.py
@@ -7,7 +7,7 @@ from django.views.generic.edit import View
 from django.utils import timezone
 
 from .definitions import JourneyDefinition, journey_list
-from .models import ActiveJourney
+from .models import ActiveJourney, toggle_journey_log_visibility
 
 
 class JourneyView(View):
@@ -30,21 +30,27 @@ class JourneyView(View):
         return render(request, 'journeys.html', template_context)
 
     def post(self, request):
-        if len(ActiveJourney.objects.filter(character=request.character, active=True)) > 0:
-            return HttpResponseServerError(status=500)
-        ActiveJourney.objects.filter(character=request.character).delete()
-        journey_id = request.POST['journey_id']
-        print(journey_id)
-        definition: JourneyDefinition = next(filter(lambda j: j.slug == journey_id, journey_list))
-        print(definition.name)
-        log = list(definition.proceed(request.character))
-        print(log)
-        duration = sum([l.duration for l in log])
+        if 'journey_id' in request.POST:
+            if len(ActiveJourney.objects.filter(character=request.character, active=True)) > 0:
+                return HttpResponseServerError(status=500)
+            ActiveJourney.objects.filter(character=request.character).delete()
+            # journey_visible = request.POST.get['is_visible']
+            journey_id = request.POST['journey_id']
+            print(journey_id)
+            definition: JourneyDefinition = next(filter(lambda j: j.slug == journey_id, journey_list))
+            print(definition.name)
+            log = list(definition.proceed(request.character))
+            print(log)
+            duration = sum([l.duration for l in log])
 
-        ActiveJourney(
-            character=request.character,
-            end_date=timezone.now() + timedelta(seconds=duration),
-            log=log,
-            slug=journey_id,
-        ).save()
+            ActiveJourney(
+                character=request.character,
+                end_date=timezone.now() + timedelta(seconds=duration),
+                log=log,
+                slug=journey_id,
+            ).save()
+        elif 'is_visible' in request.POST:
+            last_journey = ActiveJourney.objects.filter(character=request.character, active=False).first()
+            if last_journey is not None:
+                toggle_journey_log_visibility(last_journey)
         return redirect('journeys')
